fix(#602): move CROSS_ZONE_ALLOWED_RELATIONS to rebac brick
The rebac/ brick imported CROSS_ZONE_ALLOWED_RELATIONS from
services.permissions.cross_zone, creating a circular dependency
(brick → services layer). The constant belongs in rebac/ since it
defines relation semantics owned by the ReBAC brick.

- Created nexus/rebac/cross_zone.py with the canonical definition
- Updated 3 rebac/ importers (manager.py, bulk_checker.py,
  consistency/zone_manager.py) to use nexus.rebac.cross_zone
- services/permissions/cross_zone.py now re-exports from rebac
