+++ src/nexus/rebac/batch/bulk_checker.py
@@ -23,8 +23,8 @@ from typing import TYPE_CHECKING, Any
 from sqlalchemy.exc import OperationalError
 
 from nexus.core.rebac import Entity
+from nexus.rebac.cross_zone import CROSS_ZONE_ALLOWED_RELATIONS
 from nexus.rebac.types import ConsistencyLevel
-from nexus.services.permissions.cross_zone import CROSS_ZONE_ALLOWED_RELATIONS
 
 if TYPE_CHECKING:
     from collections.abc import Callable
