+++ src/nexus/rebac/consistency/zone_manager.py
@@ -25,7 +25,7 @@ from __future__ import annotations
 
 import logging
 
-from nexus.services.permissions.cross_zone import CROSS_ZONE_ALLOWED_RELATIONS
+from nexus.rebac.cross_zone import CROSS_ZONE_ALLOWED_RELATIONS
 
 logger = logging.getLogger(__name__)
